Refactor to mock App for unit testing httpServer
