+++ .vscode/launch.json
@@ -4,7 +4,6 @@
     // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
     "version": "0.2.0",
     "configurations": [
-
         {
             "type": "node",
             "request": "launch",
@@ -14,13 +13,30 @@
         {
             "type": "node",
             "request": "launch",
-            "name": "Jest Tests",
+            "name": "Jest All Old",
             "program": "${workspaceFolder}/node_modules/jest/bin/jest.js",
             "args": [
                 "--coverage",
                 "--detectOpenHandles"
             ],
             "internalConsoleOptions": "openOnSessionStart"
+        },
+        {
+            "type": "node",
+            "request": "launch",
+            "name": "Jest All",
+            "program": "${workspaceFolder}/node_modules/.bin/jest",
+            "args": [
+                "--runInBand",
+                "--coverage",
+                "--detectOpenHandles"
+            ],
+            "console": "integratedTerminal",
+            "internalConsoleOptions": "neverOpen",
+            "disableOptimisticBPs": true,
+            "windows": {
+                "program": "${workspaceFolder}/node_modules/jest/bin/jest",
+            }
         },
         {
             "type": "node",
@@ -34,7 +50,28 @@
             "type": "node",
             "request": "launch",
             "name": "Launch Quick Test",
-            "program": "${workspaceFolder}\\lib\\quickTest.js"
+            "program": "${workspaceFolder}/lib/quickTest.js",
+            "disableOptimisticBPs": true,
+            "windows": {
+                "program": "${workspaceFolder}\\lib\\quickTest.js",
+            }
+        },
+        {
+            "type": "node",
+            "request": "launch",
+            "name": "Jest Current File",
+            "program": "${workspaceFolder}/node_modules/.bin/jest",
+            "args": [
+                "${fileBasenameNoExtension}",
+                "--config",
+                "jest.config.js"
+            ],
+            "console": "integratedTerminal",
+            "internalConsoleOptions": "neverOpen",
+            "disableOptimisticBPs": true,
+            "windows": {
+                "program": "${workspaceFolder}/node_modules/jest/bin/jest",
+            }
         }
     ]
 }
