+++ package.json
@@ -17,22 +17,22 @@
     "botframework-config": "^4.3.4",
     "mongodb": "^3.2.3",
     "node-fetch": "^2.3.0",
-    "restify": "^7.7.0"
+    "restify": "^7.7.0",
+    "ts-jest": "^24.0.2"
   },
   "devDependencies": {
     "@microsoft/microsoft-graph-types": "^1.8.0",
     "@microsoft/microsoft-graph-types-beta": "github:microsoftgraph/msgraph-typescript-typings#beta",
     "@microsoft/recognizers-text-date-time": "^1.1.4",
     "@types/dotenv": "^6.1.1",
-    "@types/jest": "^24.0.11",
+    "@types/jest": "^24.0.12",
     "@types/mongodb": "^3.1.24",
     "@types/node-fetch": "^2.3.2",
     "@types/restify": "^7.2.9",
     "concurrently": "^4.0.1",
-    "dotenv": "^7.0.0",
+    "dotenv": "^8.0.0",
     "jest": "^24.8.0",
     "nodemon": "^1.18.11",
-    "ts-jest": "^24.0.2",
     "typescript": "^3.4.5"
   },
   "engines": {
