+++ src/quickTest.ts
@@ -1,10 +1,6 @@
 // Note that normally will never be called.  It is useful for a quick check on app initialization and module load debugging.
 
-import { app } from './app';
+import { app, create } from './nagbotApp';
 
-async function test() {
-    await app.ready;
-    console.log(app);
-}
+create();
 
-test();
